feat: ship landing page with mobile and desktop variants
